+++ src/FakeRobot.Infrastructure/Repositories/Interface/IRobotCommandRepository.cs
@@ -4,5 +4,5 @@ namespace FakeRobot.Infrastructure.Repositories.Interface;
 
 public interface IRobotCommandRepository
 {
-    CommandsSummary SaveRobotCommandResult(int commands, int result, double duration);
+    CommandsSummary SaveRobotCommandResult(int commands, int result, long duration);
 }
